+++ src/cli/commands/mod.rs
@@ -7,4 +7,4 @@ pub mod migrate;
 pub mod release_metadata;
 pub mod snippets;
 pub mod validate_versions;
-mod version_manifests;
+pub(crate) mod version_manifests;